+++ src/Products/PloneMeeting/interfaces.py
@@ -141,6 +141,18 @@ class IMeetingBatchActionsMarker(IBatchActionsMarker):
     """Marker interfaces to register batch actions for Meetings."""
 
 
+class IMeetingItemBatchActionsMarker(IBatchActionsMarker):
+    """Marker interfaces to register batch actions for MeetingItems."""
+
+
+class IMeetingAdviceBatchActionsMarker(IBatchActionsMarker):
+    """Marker interfaces to register batch actions for MeetingAdvices."""
+
+
+class IMeetingContentBatchActionsMarker(IBatchActionsMarker):
+    """Marker interfaces to register batch actions for every MeetingContents."""
+
+
 class IToolPloneMeeting(Interface):
     """Marker interface for .ToolPloneMeeting.ToolPloneMeeting
     """